foodtask/requirements.txt
django==5.0.7
djangorestframework==3.15.2
django_model_utils==4.5.1